+++ test/fixtures/simple-transport/c.js
@@ -1,2 +1,3 @@
 require("./a");
 require("b");
+//require("d");