+++ HISTORY.md
@@ -7,6 +7,7 @@
 - stop parsing dependencies when have been ignore
 - ignore condition before package check
 - throw when rename error
+- replace ignore comment
 
 ## 0.4.4